requirements.txt
Django==3.2.12
djangorestframework==3.13.1
django-cors-headers==3.10.1